+++ Package.swift
@@ -1,30 +1,27 @@
 // swift-tools-version:5.1
-// The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
 
 let package = Package(
     name: "Mortar",
+    platforms: [.iOS(.v12), .tvOS(.v12), .macOS(.v10_14)],
     products: [
-        // Products define the executables and libraries produced by a package, and make them visible to other packages.
         .library(
             name: "Mortar",
-            targets: ["Mortar"]),
+            targets: ["Mortar"]
+        ),
     ],
-    dependencies: [
-        // Dependencies declare other packages that this package depends on.
-        // .package(url: /* package url */, from: "1.0.0"),
-    ],
+    dependencies: [],
     targets: [
-        // Targets are the basic building blocks of a package. A target can define a module or a test suite.
-        // Targets can depend on other targets in this package, and on products in packages which this package depends on.
         .target(
             name: "Mortar",
             dependencies: [],
-            path: "Mortar"),
+            path: "Mortar"
+        ),
         .testTarget(
             name: "MortarTests",
             dependencies: ["Mortar"],
-            path: "MortarTests"),
+            path: "MortarTests"
+        ),
     ]
 )